PMA-127 - Property related policies

+++ app/Http/Controllers/PropertyLinkCategoryController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\PropertyLinkCategory\UpdateRequest;
 use App\Http\Resources\PropertyLinkCategoryResource;
 use App\Http\Resources\PropertyLinkCategoryResourceCollection;
 use App\Repositories\Contracts\PropertyLinkCategoryRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class PropertyLinkCategoryController extends Controller
 {
@@ -31,9 +32,12 @@ class PropertyLinkCategoryController extends Controller
      *
      * @param IndexRequest $request
      * @return PropertyLinkCategoryResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [PropertyLinkCategory::class, $request->get('propertyId')]);
+
         $propertyLinkCategories = $this->propertyLinkCategoryRepository->findBy($request->all());
 
         return new PropertyLinkCategoryResourceCollection($propertyLinkCategories);
@@ -44,9 +48,12 @@ class PropertyLinkCategoryController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @return PropertyLinkCategoryResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [PropertyLinkCategory::class, $request->get('propertyId')]);
+
         $propertyLinkCategory = $this->propertyLinkCategoryRepository->save($request->all());
 
         return new PropertyLinkCategoryResource($propertyLinkCategory);
@@ -57,9 +64,12 @@ class PropertyLinkCategoryController extends Controller
      *
      * @param PropertyLinkCategory $propertyLinkCategory
      * @return PropertyLinkCategoryResource
+     * @throws AuthorizationException
      */
     public function show(PropertyLinkCategory $propertyLinkCategory)
     {
+        $this->authorize('show', $propertyLinkCategory);
+
         return new PropertyLinkCategoryResource($propertyLinkCategory);
     }
 
@@ -69,9 +79,12 @@ class PropertyLinkCategoryController extends Controller
      * @param UpdateRequest $request
      * @param PropertyLinkCategory $propertyLinkCategory
      * @return PropertyLinkCategoryResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, PropertyLinkCategory $propertyLinkCategory)
     {
+        $this->authorize('update', $propertyLinkCategory);
+
         $propertyLinkCategory = $this->propertyLinkCategoryRepository->update($propertyLinkCategory, $request->all());
 
         return new PropertyLinkCategoryResource($propertyLinkCategory);
@@ -82,9 +95,12 @@ class PropertyLinkCategoryController extends Controller
      *
      * @param PropertyLinkCategory $propertyLinkCategory
      * @return \Illuminate\Http\Response
+     * @throws AuthorizationException
      */
     public function destroy(PropertyLinkCategory $propertyLinkCategory)
     {
+        $this->authorize('destroy', $propertyLinkCategory);
+
         $this->propertyLinkCategoryRepository->delete($propertyLinkCategory);
 
         return response()->json(null, 204);
